+++ extensions/Worker.Extensions.EventGrid/src/release_notes.md
@@ -1,3 +1,3 @@
 ### Microsoft.Azure.Functions.Worker.Extensions.EventGrid <version>
 
-- Add ability to bind a event grid trigger to CloudEvent, CloudEvent[], EventGridEvent, EventGridEvent[], BinaryData, BinaryData[], and String[]
+- Add ability to bind a event grid trigger to CloudEvent, CloudEvent[], EventGridEvent, EventGridEvent[], BinaryData, BinaryData[], and string[]